chore: use private fields in the container
