+++ modules/etcd/etcd_test.go
@@ -2,7 +2,6 @@ package etcd_test
 
 import (
 	"context"
-	"fmt"
 	"io"
 	"testing"
 	"time"
@@ -39,35 +38,8 @@ func TestRunCluster1Node(t *testing.T) {
 	require.NoError(t, err)
 
 	// the topology has only one node with no children
-	require.Empty(t, ctr.Nodes)
+	require.Equal(t, 0, ctr.NodesCount())
 	require.Equal(t, etcd.DefaultClusterToken, ctr.ClusterToken)
-}
-
-func TestRunClusterMultipleNodes(t *testing.T) {
-	ctx := context.Background()
-
-	ctr, err := etcd.Run(ctx, "gcr.io/etcd-development/etcd:v3.5.14", etcd.WithNodes("etcd-1", "etcd-2", "etcd-3"), etcd.WithClusterToken("My-cluster-t0k3n"))
-	testcontainers.CleanupContainer(t, ctr)
-	require.NoError(t, err)
-
-	require.Equal(t, "My-cluster-t0k3n", ctr.ClusterToken)
-
-	// the topology has one parent node and two child nodes
-	require.Len(t, ctr.Nodes, 2)
-
-	for i, node := range ctr.Nodes {
-		require.Empty(t, node.Nodes) // child nodes has no children
-
-		c, r, err := node.Exec(ctx, []string{"etcdctl", "member", "list"}, tcexec.Multiplexed())
-		require.NoError(t, err)
-
-		output, err := io.ReadAll(r)
-		require.NoError(t, err)
-		require.Contains(t, string(output), fmt.Sprintf("etcd-%d", i+1))
-
-		require.Equal(t, 0, c)
-		require.Equal(t, "My-cluster-t0k3n", node.ClusterToken)
-	}
 }
 
 func TestRun_PutGet(t *testing.T) {
